+++ Classes/ViewHelpers/GetTotalNumberOfPagesAliasViewHelper.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace Bithost\Pdfviewhelpers\Model;
+namespace Bithost\Pdfviewhelpers\ViewHelpers;
 
-/***
+/* * *
  *
  * This file is part of the "PDF ViewHelpers" Extension for TYPO3 CMS.
  *
@@ -26,36 +26,20 @@ namespace Bithost\Pdfviewhelpers\Model;
  *  GNU General Public License for more details.
  *
  *  This copyright notice MUST APPEAR in all copies of the script!
- ***/
-
-use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
+ * * */
 
 /**
- * BithostTCPDF
+ * GetTotalNumberOfPagesAliasViewHelper
  *
  * @author Markus Mächler <markus.maechler@bithost.ch>, Esteban Marin <esteban.marin@bithost.ch>
  */
-class BithostTCPDF extends \TCPDF
+class GetTotalNumberOfPagesAliasViewHelper extends AbstractPDFViewHelper
 {
     /**
-     * @return void
+     * @return string
      */
-    public function Header() // phpcs:ignore
+    public function render()
     {
-        $extPath = ExtensionManagementUtility::extPath('pdfviewhelpers');
-        $address = "Bithost GmbH \nMilchbuckstrasse 83 \nCH-8057 Zürich \n\nhallo@bithost.ch \n044 585 28 20 \n\nwww.bithost.ch";
-
-        $this->SetTextColor(140, 140, 140);
-        $this->SetFontSize(11);
-
-        $this->Image($extPath . 'Resources/Public/Examples/BasicUsage/logo.png', 15, 15, 56, 24, '', '', '', false, 300, '', false, false, 0, false, false, false, false);
-        $this->MultiCell(null, null, $address, 0, 'R', false, 1, 0, 45, true, 0, false, true, 0, 'T', false);
-    }
-
-    /**
-     * @return void
-     */
-    public function Footer() // phpcs:ignore
-    {
+        return $this->getPDF()->getAliasNbPages();
     }
 }