+++ Tests/Functional/Fixtures/Examples/ExtendExistingPDFsTest.php
@@ -28,14 +28,14 @@ namespace Bithost\Pdfviewhelpers\Tests\Functional\Fixtures\Examples;
  *  This copyright notice MUST APPEAR in all copies of the script!
  ***/
 
-use Bithost\Pdfviewhelpers\Model\EmptyFPDI;
+use Bithost\Pdfviewhelpers\Model\BasePDF;
 
 /**
  * FullFeatureShowCaseTest
  *
  * @author Markus Mächler <markus.maechler@bithost.ch>, Esteban Marin <esteban.marin@bithost.ch>
  */
-class ExtendExistingPDFsTest extends EmptyFPDI
+class ExtendExistingPDFsTest extends BasePDF
 {
     /**
      * @return void